# docs/OPENAPI_FULL.yaml
openapi: 3.0.3
info:
  title: SMIF Gestão API — Full
  version: 0.3.0
  description: Especificação completa (Auth, Admin, RBAC, Localities, Specialties,
    Meetings, Tasks, Reports, Notices, Checklists, KPIs, Org/Elos, Dashboards).
servers:
- url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        message:
          type: string
        code:
          type: string
        details: {}
      required:
      - message
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
      - email
      - password
    TokenResponse:
      type: object
      properties:
        accessToken:
          type: string
      required:
      - accessToken
    LoginResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
        user:
          $ref: '#/components/schemas/UserSafe'
      required:
      - accessToken
      - refreshToken
      - user
    UserSafe:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        role:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
      required:
      - id
      - name
      - email
...
    Role:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        isSystemRole:
          type: boolean
      required:
      - id
      - name
...
    Specialty:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        color:
          type: string
          nullable: true
        icon:
          type: string
          nullable: true
      required:
      - id
      - name
    Locality:
      type: object
      properties:
        id:
          type: string
        code:
          type: string
        name:
          type: string
        commandName:
          type: string
          nullable: true
        commanderName:
          type: string
          nullable: true
        visitDate:
          type: string
          format: date-time
          nullable: true
        recruitsFemaleCountCurrent:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required:
      - id
      - code
      - name
...
    Meeting:
      type: object
      properties:
        id:
          type: string
        datetime:
          type: string
          format: date-time
        scope:
          type: string
          enum:
          - NATIONAL
          - LOCALITY
        status:
          type: string
          enum:
          - PLANNED
          - HELD
          - CANCELLED
        agenda:
          type: string
          nullable: true
        localityId:
          type: string
          nullable: true
        participantsJson: {}
      required:
      - id
      - datetime
      - scope
      - status
...
    Checklist:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        phaseId:
          type: string
          nullable: true
        specialtyId:
          type: string
          nullable: true
      required:
      - id
      - title
...
security:
- bearerAuth: []
paths:
  /auth/login:
    post:
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/refresh:
    post:
      tags:
      - Auth
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
              required:
              - refreshToken
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenResponse'
...
  /roles:
    get:
      tags:
      - RBAC
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Role'
    post:
      tags:
      - RBAC
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '201':
          description: Created
  /roles/{id}:
    put:
      tags:
      - RBAC
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Role'
      responses:
        '200':
          description: OK
    delete:
      tags:
      - RBAC
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
  /roles/{id}/clone:
    post:
      tags:
      - RBAC
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Created
...
  /roles/{id}/permissions:
    put:
      tags:
      - RBAC
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                permissions:
                  type: array
                  items:
                    type: object
              required:
              - permissions
      responses:
        '200':
          description: OK
...
  /localities:
    get:
      tags:
      - Localities
      responses:
        '200':
          description: OK
    post:
      tags:
      - Localities
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Locality'
      responses:
        '201':
          description: Created
  /localities/{id}:
    put:
      tags:
      - Localities
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Locality'
      responses:
        '200':
          description: OK
    delete:
      tags:
      - Localities
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
  /specialties:
    get:
      tags:
      - Specialties
      responses:
        '200':
          description: OK
    post:
      tags:
      - Specialties
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specialty'
      responses:
        '201':
          description: Created
  /specialties/{id}:
    put:
      tags:
      - Specialties
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Specialty'
      responses:
        '200':
          description: OK
    delete:
      tags:
      - Specialties
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '204':
          description: No Content
...
  /meetings:
    get:
      tags:
      - Meetings
      responses:
        '200':
          description: OK
    post:
      tags:
      - Meetings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meeting'
      responses:
        '201':
          description: Created
  /meetings/{id}:
    put:
      tags:
      - Meetings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Meeting'
      responses:
        '200':
          description: OK
  /meetings/{id}/decisions:
    post:
      tags:
      - Meetings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                text:
                  type: string
              required:
              - text
      responses:
        '201':
          description: Created
  /meetings/{id}/generate-tasks:
    post:
      tags:
      - Meetings
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
...
  /reports/upload:
    post:
      tags:
      - Reports
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                taskInstanceId:
                  type: string
              required:
              - file
              - taskInstanceId
      responses:
        '201':
          description: Created
  /reports/{id}/download:
    get:
      tags:
      - Reports
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        '200':
          description: OK
  /reports/{id}/approve:
    put:
      tags:
      - Reports
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                approved:
                  type: boolean
              required:
              - approved
      responses:
        '200':
          description: OK
...
  /checklists:
    get:
      tags:
      - Checklists
      responses:
        '200':
          description: OK
    post:
      tags:
      - Checklists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Checklist'
      responses:
        '201':
          description: Created
  /checklists/{id}/items:
    post:
      tags:
      - Checklists
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '201':
          description: Created
  /checklist-item-status/batch:
    put:
      tags:
      - Checklists
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: OK
...